Successfully Seperated most components
